pyproject.toml
[tool.poetry]
name = "blog_server"
version = "0.1.0"
description = ""
authors = ["vlzh <vlzh@pm.me>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.0.6"
djangorestframework = "^3.11.0"
psycopg2-binary = "^2.8.5"
python-dotenv = "^0.13.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.8.2"
python-language-server = "^0.33.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
